requirements.txt
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
lxml==4.6.2
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.4
six==1.15.0
xl2dict==0.1.5
xlrd==1.2.0
XlsxWriter==1.3.7